[REFACTOR] #42 지원자뷰 사이드바 폴더구조 리팩토링
