+++ src/pages/interviewApply/end/InterviewApplyEnd.tsx
@@ -1,8 +1,7 @@
 import { useNavigate } from 'react-router-dom';
-import * as I from './InterviewApply.style';
+import * as I from './InterviewApplyEnd.style';
 
-function InterviewApplyYet() {
-
+function InterviewApplyEnd() {
 	const navigate = useNavigate();
 
 	//회원가입 클릭 시
@@ -11,11 +10,13 @@ function InterviewApplyYet() {
 	};
 
 	return (
-		<>	
-            {/*Main*/}
+		<>
+			{/*Main*/}
 			<I.InterviewMain>
 				<I.InterviewMainText>
-					아직 지원기간이 아닙니다!
+					지원기간이 마감되었습니다.
+					<br />
+					감사합니다 :)
 				</I.InterviewMainText>
 				<I.ApplyBtnWrap>
 					<I.ApplyBtn onClick={onClickProfile}>프로필 페이지로 이동</I.ApplyBtn>
@@ -25,4 +26,4 @@ function InterviewApplyYet() {
 	);
 }
 
-export default InterviewApplyYet;
+export default InterviewApplyEnd;
