+++ src/pages/interviewApply/yet/InterviewApplyYet.tsx
@@ -1,8 +1,7 @@
 import { useNavigate } from 'react-router-dom';
-import * as I from './InterviewApply.style';
+import * as I from './InterviewApplyYet.style';
 
-function InterviewApplyEnd() {
-
+function InterviewApplyYet() {
 	const navigate = useNavigate();
 
 	//회원가입 클릭 시
@@ -11,13 +10,10 @@ function InterviewApplyEnd() {
 	};
 
 	return (
-		<>	
-            {/*Main*/}
+		<>
+			{/*Main*/}
 			<I.InterviewMain>
-				<I.InterviewMainText>
-					지원기간이 마감되었습니다.<br/>
-					감사합니다 :)
-				</I.InterviewMainText>
+				<I.InterviewMainText>아직 지원기간이 아닙니다!</I.InterviewMainText>
 				<I.ApplyBtnWrap>
 					<I.ApplyBtn onClick={onClickProfile}>프로필 페이지로 이동</I.ApplyBtn>
 				</I.ApplyBtnWrap>
@@ -26,4 +22,4 @@ function InterviewApplyEnd() {
 	);
 }
 
-export default InterviewApplyEnd;
+export default InterviewApplyYet;
